rename readline to readfile
