+++ filebeat/reader/readfile/encoding/encoding.go
