+++ filebeat/reader/readfile/encoding/mixed.go
