+++ filebeat/reader/readfile/encoding/utf16_test.go
